fixed param file

+++ maps/maps_for_jasper.sh
@@ -29,7 +29,7 @@ trunc_type="_new_truncate.mtz"
 res_high=1.5 #Highest resolution term eg 1.3
 res_low=30 #Lowest resolution term eg 15
 SYMM=19 # Symmetry group eg P212121 or 19
-
+clean="yes" #if you do not want to keep all the intermediate files
 if [ ! -f $dark_model ]; then echo "$dark_model not found!" && exit 1 ; fi
 if [ ! -f $dark_obs ]; then echo "$dark_obs not found!" && exit 1; fi
 if [ ! -d $loc ]; then echo "$loc not found!" && exit 1 ; fi
@@ -126,17 +126,29 @@ do
   ${loc}/trunc_all_options.sh $f $res_high $res_low "$cell" || exit
   ${loc}/map_generate.sh $dark_model $dark_FC $dark_OBS ./${name}${trunc_type} $name $res_high $res_low "$cell" >> map.log || exit
   ${loc}/own_scales/make_phenix_q_weight.sh $dark_model $dark_FC $dark_OBS ./${name}${trunc_type} "$cell" $res_high $res_low|| exit
-  ${loc}/Extrap_maps_marius.sh "$cell" $res_high $res_low || exit
+  ${loc}/extrap_maps_marius.sh "$cell" $res_high $res_low || exit
 
   ######################################
-  ## Fit the extraolated maps
+  ## Clean up is the clean option is y
   ######################################
-  phenix.python ${loc}/N_ext_fit_conv.py neg_count_Mari.dat
-  mv N_ext_Fitted_C.png count_MS.png
-  phenix.python ${loc}/N_ext_fit_conv_2.py neg_unw_count_Phenix.dat
-  mv N_ext_Fitted_C.png count_unw.png
-  phenix.python ${loc}/N_ext_fit_conv_2.py neg_count_Phenix.dat
-  mv N_ext_Fitted_C.png count_JB.png
+  if [ "$clean" == "yes" ] ; then
+   rm Fext_QW_SCALIT*.mtz Fext_QW_SCALIT*.map
+  fi
+
+  ######################################
+  ## Fit the extraolated maps and gernate only the Next value calaulated
+  ######################################
+  Next1=$(phenix.python ${loc}/N_ext_fit_conv_2.py neg_count_Mari.dat | awk '{print $3}')
+  mv N_ext_Fitted_C.png count_QW_Scaleit.png
+  ${loc}/extrap_MS_call.sh "$cell" $Next1
+
+  Next2=$(phenix.python ${loc}/N_ext_fit_conv_2.py neg_unw_count_Phenix.dat | awk '{print $3}')
+  mv N_ext_Fitted_C.png count_unw_Phenix_Multiscale.png
+  ${loc}/own_scales/make_single_multiscale.sh "$cell" $Next2
+
+  Next3=$(phenix.python ${loc}/N_ext_fit_conv_2.py neg_count_Phenix.dat | awk'{print $3}')
+  mv N_ext_Fitted_C.png count_QW_Phenix_multiscale.png
+  ${loc}/own_scales/make_single_qw_multiscale.sh "$cell" $Next3
 #  a=$(phenix.python ${loc}/N_ext_fit_conv_2.py neg_unw_count.dat | awk '{print int($3)}')
 #  b=$(($a*10+1))
   ######################################
@@ -150,9 +162,11 @@ do
   # Clean up
   ######################################
   params > params.dat
-  rm -f all_sc1.mtz all_sc2_free.mtz all_sc2.mtz dark_phase.hkl dark_scaled.hkl FC_dark.mtz Light_dark.phs Light_dwt.mtz light_scaled.hkl Light_wd.map model_phs.hkl scale_it.mtz wmar.inp
-  rm all_Fs.hkl all.mtz Fext_map.dat Fext_map.hkl fmodel fobs_dark fobs_light Fxt_map_9.map James_ext.mtz LFs_tmp1  light_Fs light_Fs_scaled light_Fs_scaled.hkl light.hkl neg_9.log  neg.inp neg_int.dat phenix_Fobs_Fobs.mtz tmp_dark_1 tmp_dark_12a tmp_dark_1a tmp_dark_2 tmp_dark_2a tmp_dark_phase_1  tmp_dark_phase_12a tmp_dark_phase_1a  tmp_dark_phase_2 tmp_dark_phase_2a tmp_light_1  tmp_light_12a tmp_light_1a tmp_light_2 tmp_light_2a weighted_map_log
-  rm dark_Fs dark_Fs_scaled dark_Fs_scaled.hkl dark.hkl  dark_phases dark_phases.hkl dark_phases_sorted.hkl dFC_DFs_LFs_tmp1 dFC_DFs_LFs_tmp2 dFC_DFs_tmp1  dFC_tmp1 DFs_tmp1 diff_unweighted_map.dat diff_unweighted_map.hkl diff_unweighted_map.map diff_unweighted_map.mtz diff_weight_map.dat diff_weight_map.hkl diff_weight_map.mtz map_mol.ccp4 py_tmp.pml neg_*.log
 
+  if [ "$clean" == "yes" ] ; then
+    rm -f all_sc1.mtz all_sc2_free.mtz all_sc2.mtz dark_phase.hkl dark_scaled.hkl FC_dark.mtz Light_dark.phs Light_dwt.mtz light_scaled.hkl Light_wd.map model_phs.hkl scale_it.mtz wmar.inp
+    rm all_Fs.hkl all.mtz Fext_map.dat Fext_map.hkl fmodel fobs_dark fobs_light Fxt_map_9.map James_ext.mtz LFs_tmp1  light_Fs light_Fs_scaled light_Fs_scaled.hkl light.hkl neg_9.log  neg.inp neg_int.dat phenix_Fobs_Fobs.mtz tmp_dark_1 tmp_dark_12a tmp_dark_1a tmp_dark_2 tmp_dark_2a tmp_dark_phase_1  tmp_dark_phase_12a tmp_dark_phase_1a  tmp_dark_phase_2 tmp_dark_phase_2a tmp_light_1  tmp_light_12a tmp_light_1a tmp_light_2 tmp_light_2a weighted_map_log
+    rm dark_Fs dark_Fs_scaled dark_Fs_scaled.hkl dark.hkl  dark_phases dark_phases.hkl dark_phases_sorted.hkl dFC_DFs_LFs_tmp1 dFC_DFs_LFs_tmp2 dFC_DFs_tmp1  dFC_tmp1 DFs_tmp1 diff_unweighted_map.dat diff_unweighted_map.hkl diff_unweighted_map.map diff_unweighted_map.mtz diff_weight_map.dat diff_weight_map.hkl diff_weight_map.mtz map_mol.ccp4 py_tmp.pml neg_*.log
+  fi
 done
